requirements.txt
Django==5.2.8
djangorestframework
djangorestframework-simplejwt
pandas
django-cors-headers